+++ CHANGELOG.md
@@ -8,10 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upgraded Node to version 16.x
 
+### Breaking changes
+- Flow step type `media` is now `singleMedia` and takes field `acceptOnly` to assert accepted media type
+- Responses to steps of type `singleMedia` are now saved as array of strings
+
 ### Added
+- Bot now responds to messages of unsupported type with an appropriate error message
 - [#3](https://github.com/opengeolab/geocollectorbot/issues/3): exposed `POST - /send-message` route to programmatically send messaged to chats
+- [#6](https://github.com/opengeolab/geocollectorbot/issues/6): added support for video messages
 - [#7](https://github.com/opengeolab/geocollectorbot/issues/7): exposed `GET - /interactions` route to return all saved interactions
-- Bot now responds to messages of unsupported type with an appropriate error message 
 
 ## [1.0.0] 2022-03-12
 - First release